moving tests around for flash_attn install
